refactor: api 관련 파일 TS 마이그레이션 및 fetcher 함수 추상화
