+++ src/domain/movieService.ts
@@ -1,19 +1,4 @@
-export interface MovieResponse {
-  adult: boolean;
-  backdrop_path: string;
-  genre_ids: number[];
-  id: number;
-  original_language: string;
-  original_title: string;
-  overview: string;
-  popularity: number;
-  poster_path: string;
-  release_date: string;
-  title: string;
-  video: boolean;
-  vote_average: number;
-  vote_count: number;
-}
+import { MovieResponse } from '../api';
 
 interface Movie {
   title: MovieResponse['title'];
